Use native configuration option for IPMI sessions
This paves the road for switching to using an IPMI library instead of
parsing ipmitool output. The old remote session configuration format of
specifying ipmitool arguments is still supported, but only if the
arguments are `-I lanplus`, `-H`, `-U`, or `-P`. The backwards
compatibility layer will not support other arguments.

Issue: #50

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>